Add an `analyse:;array_check` module which checks for malformed arrays and/or indexing.
Also pulled the type checker into its own module.
