+++ pintc/tests/arrays/compare_different_sizes.pnt
@@ -1,4 +1,3 @@
-// <disabled>
 var a: int[4];
 var b = [11, 22, 33, 44, 55];
 var c = [66, 77];
