+++ pintc/tests/arrays/non_const_index.pnt
@@ -1,4 +1,3 @@
-// <disabled>
 var n: int;
 var a: int[10];
