+++ pintc/tests/arrays/non_const_length.pnt
@@ -1,4 +1,3 @@
-// <disabled>
 var n: int;
 var a: int[n];
