+++ pintc/tests/arrays/out_of_bounds_index.pnt
@@ -1,4 +1,3 @@
-// <disabled>
 var b: int[4];
 
 constraint b[4] == 2;
