+++ pintc/tests/arrays/out_of_bounds_index_multidimensional_second.pnt
@@ -1,4 +1,3 @@
-// <disabled>
 var b: int[4][2];
 
 constraint b[1][3] == 2;
